requirements.txt
# OSI News Automation System - Requirements
# Python 3.11.9 | Windows Compatible
# ============================================

# ===========================================
# Web Scraping
# ===========================================
newspaper3k==0.2.8
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
playwright==1.40.0

# ===========================================
# NLP & ML
# ===========================================
transformers==4.36.0
sentence-transformers==2.2.2
torch==2.1.2
langdetect==1.0.9

# ===========================================
# LLM Integration (FREE - Groq)
# ===========================================
groq==0.4.1

# ===========================================
# Image Generation (Local Stable Diffusion)
# ===========================================
diffusers==0.25.0
accelerate==0.25.0
safetensors==0.4.1
pillow==10.1.0

# ===========================================
# Database (MongoDB)
# ===========================================
pymongo==4.6.1
motor==3.3.2  # Async MongoDB driver

# ===========================================
# Translation
# ===========================================
deep-translator==1.11.4

# ===========================================
# Scheduling & Automation
# ===========================================
APScheduler==3.10.4
python-crontab==3.0.0

# ===========================================
# API Integrations (Social Media)
# ===========================================
tweepy==4.14.0        # Twitter
python-linkedin==4.1  # LinkedIn
instagrapi==2.0.0     # Instagram
facebook-sdk==3.1.0   # Facebook

# ===========================================
# Utilities
# ===========================================
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3
validators==0.22.0

# ===========================================
# Logging & Monitoring
# ===========================================
loguru==0.7.2

# ===========================================
# Testing
# ===========================================
pytest==7.4.3
pytest-asyncio==0.21.1
